+++ wrapper/Tools/FreeEnergyAnalysis.py
@@ -26,7 +26,7 @@ except ImportError:
 try:
     # This should force the installation of pymbar if it isn't
     # installed already
-    _pymbar = Sire.try_import("pymbar", version="<4")
+    _pymbar = Sire.try_import("pymbar", version=">=4")
 
     from pymbar import MBAR
     from pymbar import timeseries
@@ -105,21 +105,16 @@ class FreeEnergies(object):
         try:
             MBAR_obj = MBAR(self._u_kln, self._N_k, verbose=True)
             self._f_k = MBAR_obj.f_k
-            (
-                deltaF_ij,
-                dDeltaF_ij,
-                theta_ij,
-            ) = MBAR_obj.getFreeEnergyDifferences(return_theta=True)
+            results = MBAR_obj.compute_free_energy_differences()
+            deltaF_ij = results["Delta_f"]
+            dDeltaF_ij = results["dDelta_f"]
         except:
             solver_options = {"maximum_iterations": 10000, "verbose": True}
             solver_protocol = {"method": "BFGS", "options": solver_options}
             MBAR_obj = MBAR(self._u_kln, self._N_k, solver_protocol=(solver_protocol,))
-            self._f_k = MBAR_obj.f_k
-            (
-                deltaF_ij,
-                dDeltaF_ij,
-                theta_ij,
-            ) = MBAR_obj.getFreeEnergyDifferences(return_theta=True)
+            results = MBAR_obj.compute_free_energy_differences()
+            deltaF_ij = results["Delta_f"]
+            dDeltaF_ij = results["dDelta_f"]
         self._deltaF_mbar = deltaF_ij[0, self._lambda_array.shape[0] - 1]
         self._dDeltaF_mbar = dDeltaF_ij[0, self._lambda_array.shape[0] - 1]
         self._pmf_mbar = numpy.full(
@@ -138,7 +133,7 @@ class FreeEnergies(object):
 
         ##testing data overlap:
         if test_overlap:
-            overlap_matrix = MBAR_obj.computeOverlap()
+            overlap_matrix = MBAR_obj.compute_overlap()
             self._overlap_matrix = overlap_matrix["matrix"]
 
     @property
@@ -255,13 +250,13 @@ class SubSample(object):
                 shape=(self._gradients_kn.shape[0]), fill_value=numpy.nan
             )
             for i in range(g_k.shape[0]):
-                g_k[i] = timeseries.statisticalInefficiency(self._gradients_kn[i, :])
+                g_k[i] = timeseries.statistical_inefficiency(self._gradients_kn[i, :])
             g = int(numpy.max(g_k))
             # now we need to figure out what the indices in the data are for subsampling
             indices_k = []
             for i in range(g_k.shape[0]):
                 indices_k.append(
-                    timeseries.subsampleCorrelatedData(self._gradients_kn[i, :], g=g)
+                    timeseries.subsample_correlated_data(self._gradients_kn[i, :], g=g)
                 )
                 self._subsampled_N_k_gradients[i] = len(indices_k[i])
             N_max = int(numpy.max(self._subsampled_N_k_gradients))
@@ -319,7 +314,7 @@ class SubSample(object):
             # first we compute statistical inefficiency
             g_k = numpy.full(shape=(self._energies_kn.shape[0]), fill_value=numpy.nan)
             for i in range(g_k.shape[0]):
-                g_k[i] = timeseries.statisticalInefficiency(
+                g_k[i] = timeseries.statistical_inefficiency(
                     self._energies_kn[i, percentage_removal[i] :]
                 )
             g = numpy.max(g_k)
@@ -330,7 +325,7 @@ class SubSample(object):
             )
             for i in range(g_k.shape[0]):
                 indices_k.append(
-                    timeseries.subsampleCorrelatedData(self._energies_kn[i, :], g=g)
+                    timeseries.subsample_correlated_data(self._energies_kn[i, :], g=g)
                 )
                 self._subsampled_N_k_energies[i] = len(indices_k[i])
             # self._subsampled_N_k_energies = (numpy.ceil(self._N_k / g)).astype(int)